Create Hackerearth-Providing Gifts.cpp
ou are given  coins whose values are from  to . The value of each coin ranges from  to . You want to gift as many people as you can.

You also want to be fair with everyone and therefore, you prepare the gift such that each person can be gifted the same amount of money. You are allowed to put several denominations together in a gift. What is the maximum number of gifts that you can prepare?

Note: It does not make sense to prepare empty gifts.

Input format

The first line contains  denoting the number of test cases.
The first and only line of each test case contains an integer  as described in the problem statement.
Output format

For each test case, print a single integer consisting of the maximum number of gifts.