Create Hackerearth-Replace the strings.py
You are given two strings  and  that consist of lower case Latin letters and both strings contain single '?'. You need to replace '?' with some lower case Latin letter in both strings not necessarily the same.

After that, you can apply the following operation on  any number of times:

Swap any two adjacent characters.
Your task is to determine if it is possible to make  equal to  if you replace '?' with a lower case Latin letter and apply the mentioned operation on  any number of times.